Update errors, break up push notification class, fix typo, update project filters.
